refactor: 입력 유효성 검사 로직을 전체 재사용 하도록 개선
step2 에서도 기존 domain 디렉터리에 있던 Validation 객체를 재사용 하도록 개선
