+++ src/step2-js/LottoGameController.js
@@ -1,6 +1,6 @@
 import LottoGame from '../domain/LottoGame.js';
 import View from '../view/view.js';
-import Validation from './Validation';
+import Validation from '../domain/Validation.js';
 import { LOTTO_CONDITION } from '../constants/condition.js';
 import { $ } from '../utils/dom.js';
 
@@ -19,14 +19,7 @@ export default class LottoGameController {
 
   #createLotto(e) {
     e.preventDefault();
-
-    const purchaseAmount = $('.purchase-amount-input').value;
-    const { isValid, message } = Validation.validatePurchaseAmount(purchaseAmount);
-
-    if (!isValid) {
-      this.#view.showAlert(message);
-      return;
-    }
+    const purchaseAmount = this.#validatePurchaseAmount();
 
     const lottoQuantity = purchaseAmount / LOTTO_CONDITION.lottoPrice;
 
@@ -36,5 +29,19 @@ export default class LottoGameController {
 
       return lottoNumbers;
     });
+  }
+
+  #validatePurchaseAmount() {
+    const purchaseAmount = $('.purchase-amount-input').value;
+
+    try {
+      Validation.validatePurchaseAmount(purchaseAmount);
+
+      return Number(purchaseAmount);
+    } catch ({ message }) {
+      this.#view.showAlert(message);
+
+      return;
+    }
   }
 }
